blogsync: add token generation command
Write As has a rather odd API. Instead of generating tokens via an
OAuth2 flow, a web UI, or some other common method you first use an API
endpoint that authenticates with your username and password to get a
token, then use that token to authenticate with other API endpoints.
Instead of always logging in with the username and password, provide a
command to generate tokens that can be saved and used or revoked later.
This way security concious users can do this once on a machine that
stores secrets and then distribute the token wherever it is needed.