internal/blog: don't parse bool strings
The original idea of supporting bools or strings that can be parsed into
a bool was to support frontmatter formats that may not have a bool type.
However, I don't know if these exist and this kind of premature
optimization is exactly how we get stuck having to support lots of extra
code that no one uses later.
A better approach would be to leave this out, then if it ever becomes a
problem, make sure that the convert subcommand knows how to convert
strings to bools when it's converting the other metadata format to TOML.