blogsync: make config file params match Go struct
This means that they are accessed from a template using the same name as
actually exists in the config file. For example, previously to access
"params" in the config file you had to write "Config.Params", which is
confusing and makes it difficult to know what should and should not be
capitalized.