blogsync: convert markdown to write.as markdown
Write.as uses a weird version of markdown that won't work with
hard-wrapped files. When we're uploading, parse the markdown ourselves
and attempt to unwrap text nodes so that they don't result in blog posts
being peppered in unintentional newlines.
This sort of thing should be handled at the editor layer, not at the
markdown layer but write.as doesn't see it that way so work around it
for now.