blogsync: create collections when previewing
Previously posts in collections other than the main site would fail to
publish to the preview instance because the collection didn't exist.
This patch allows creating collectinos that don't exist during
publishing (but only if we're in preview mode, we don't want every typo
to a collection name resulting in a new collection in production).