blogsync: launch browser after publishing preview
This required some extra work into how waiting on writefreely works
because when opening a browser tab we may get a SIGCHLD back as soon as
the process we spawned to open the tab exists (if it decides that the
tab should be opened in an existing, non-child process, for instance and
then terminates itself).
We are now waiting on the writefreely process to exit in a blocked
Goroutine instead of waiting on the signall but this is fine and
probably how I should have done things in the first place as it makes
everything much less error prone.