blogsync: remove duplicate WF runner code
Previously there were two versions of the code that shells out to
writefreely. One version was for the long running process that serves
the files, the other was for shorter lived commands like generating
keys. The difference was that the latter bracketed their output, but
this is not strictly necessary and nothing is running concurrently right
now that would interrupt that output anyways, so use a single function
for this.