blogsync: add naive filesystem watching
This implementation currently deletes all posts and re-uploads them any
time a source file changes. This is naive, slow, and requires a *lot* of
API requests, but for now it works to show the concept.
The next step will be to figure out the event type and the old and new
slugs and only change the files and posts that need to be tweaked.

Fixes #4